requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
PyPDF2>=3.0.1
numpy>=1.26.0
